# app/Lib/Data/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cp Chat",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/API_HOST/app/group/join": {
      "post": {
        "summary": "申请加群",
        "deprecated": false,
        "description": "用户在聊天列表以群号加群或在大厅申请加群",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer ",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_code": {
                    "type": "string",
                    "description": "群号"
                  }
                },
                "required": [
                  "group_code"
                ]
              },
              "example": {
                "group_code": "88776"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "状态码"
                    },
                    "msg": {
                      "type": "string",
                      "description": "讯息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "group_id": {
                          "type": "integer",
                          "description": "加入的群ID"
                        }
                      },
                      "required": [
                        "group_id"
                      ],
                      "description": "回传数据"
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "01K30C8K9NBYQ5T4TKDCBK75JM"
                  ]
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "code": 0,
                      "msg": "加入成功",
                      "data": {
                        "group_id": 2
                      }
                    }
                  },
                  "2": {
                    "summary": "Group Not Found",
                    "value": {
                      "code": 127,
                      "msg": "未找到该群组"
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/API_HOST/app/group/quit": {
      "post": {
        "summary": "申请退群",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer 002B666780b0300d51db9d948b2b3d0c8d534e11754294467041",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_code": {
                    "type": "string",
                    "description": "群号"
                  }
                },
                "required": [
                  "group_code"
                ]
              },
              "example": {
                "group_code": "544222"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "code": 0,
                      "msg": "ok"
                    }
                  },
                  "2": {
                    "summary": "群组不存在",
                    "value": {
                      "code": 127,
                      "msg": "未找到该群组"
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/API_HOST/app/group/pin": {
      "post": {
        "summary": "群置顶",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer 002B666ebf06c9552aaa552f552839d2d1313521755573086706",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_id": {
                    "type": "integer",
                    "description": "群ID"
                  }
                },
                "required": [
                  "group_id"
                ]
              },
              "example": {
                "group_id": 2
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 0,
                  "msg": "置顶成功"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/API_HOST/app/group/unpin": {
      "post": {
        "summary": "取消群置顶",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer 002B6667742dd34f8f0c6340013ec95d3a7ed831755572676191",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_id": {
                    "type": "integer",
                    "description": "群ID"
                  }
                },
                "required": [
                  "group_id"
                ]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 0,
                  "msg": "取消置顶成功"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/API_HOST/app/group/info": {
      "get": {
        "summary": "群信息",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "group_id",
            "in": "query",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer 002B6667742dd34f8f0c6340013ec95d3a7ed831755572676191",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_id": {
                    "type": "integer",
                    "description": "群ID"
                  }
                },
                "required": [
                  "group_id"
                ]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "group_id": {
                          "type": "integer",
                          "description": "群ID"
                        },
                        "code": {
                          "type": "string",
                          "description": "群号"
                        },
                        "title": {
                          "type": "string",
                          "description": "群名称"
                        },
                        "bulletin": {
                          "type": "string",
                          "description": "群公告"
                        },
                        "user_count" : {
                          "type" : "integer",
                          "description": "群人数"
                        }
                      },
                      "required": [
                        "group_id",
                        "code",
                        "title",
                        "bulletin",
                        "user_count"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "data"
                  ]
                },
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {
                    "group_id": 1,
                    "code": "000001",
                    "title": "第一间",
                    "bulletin": "群公告",
                    "user_count" : 21
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/API_HOST/app/group/clean": {
      "post": {
        "summary": "清空群对话",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer 00249f268d9d9c45542ad45a3ad310221ae",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_id": {
                    "type": "integer"
                  }
                },
                "required": [
                  "group_id"
                ]
              },
              "example": {
                "group_id": 1
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 0,
                  "msg": "操作成功"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/API_HOST/app/group/update": {
      "post": {
        "summary": "管理者修改群信息",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer 0028717bb0cdc1635903effdc899d99b029",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "group_id": {
                    "type": "integer",
                    "description": "群ID"
                  },
                  "title": {
                    "type": "string",
                    "description": "群标题, 不修改的话不要送这栏或给null"
                  },
                  "bulletin": {
                    "type": "string",
                    "description": "群公告, 不修改的话不要送这栏或给null"
                  }
                },
                "required": [
                  "group_id",
                  "title",
                  "bulletin"
                ]
              },
              "example": {
                "group_id": 1,
                "title": "修改群标题-初始群修改群标题",
                "bulletin": "公告V2"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string",
                          "description": "更新后标题"
                        },
                        "bulletin": {
                          "type": "string",
                          "description": "更新后公告"
                        }
                      },
                      "required": [
                        "title",
                        "bulletin"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "data"
                  ]
                },
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {
                    "title": "betman99群",
                    "bulletin": "公告V2"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {},
...
  },
  "servers": [],
  "security": []
}